t490: investigate using isync/mbsync
Investigate using isync/mbsync to backup emails locally.

https://isync.sourceforge.io

For this, I've created a ~/.mbsyncrc file that contains my IMAP
configuration to connect to my email account and allows me to run
`mbsync oliverdavies.uk` to perform the backup.

`programs.mbsync` exists in Home Manager, so I can use this if I decide
to refactor this and have Home Manager create the configuration file.