Remove blank lines from the suggestions
My Neovim configuration always adds a blank line to the end of a file
that I don't want to be included in my suggestions.

Using `grep "\S"` was working in a Terminal but this broke when adding
it to my dwm configuration.

This approach works in dwm when triggered with the keyboard shortcut.