Add dev-commit
Add `dev-commit`, a script based on https://github.com/ThePrimeagen/dev/blob/f00f31c649e0887343a0ec10cab6cb5553ba6639/env/.local/scripts/dev-commit
that is meant to be a fallback to commit and push any missed changes
that should be kept.

This is not intended to be a substitute for making manual commits and
writing detailed commit messages.

The script uses an environment variable for the list of project
directories, which is set within a Home Manager module that also creates
a systemd service and timer to run the command automatically - similar
to a cron job.

Adding this should prevent occasional issues, such as missing the
firewall configuration for my homelab which was removed accidentally and
stopped my reverse proxies for services such as Jellyfin and Immich from
running.