Serve the new podcast episode from Tome
Now the Podcast Episode content type and the first episode node has been
created, serve the new episode from Tome.

It won't appear on the episode listing page at /podcast, but I'll
recreate that once I've migrated the existing episodes to Tome, or I can
manually add it to the listing page by editing the HTML rather than
re-generating the page with Sculpin as it would confuse the commit
history in this project - which I want to keep intact.